Remove `PlannerModelManager`
`PlannerModelManager` has been removed. Its methods have been moved to
`ModelManager`.

Fixed minor bug where the student name was used instead of the major name in
`Student#toString`.
