+++ src/main/java/seedu/address/logic/commands/ModuleAddCommand.java
@@ -3,11 +3,8 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.PlannerModelManager;
 import seedu.address.model.module.ModuleCode;
 import seedu.address.model.student.Enrollment;
-import seedu.address.model.student.Major;
-import seedu.address.model.student.Student;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
